getTBinR: Get WHO Tuberculosis Data in R

The World Health Organisation compiles multiple datasets across infectious diseases. This package aims to facilitate easy import of analysis ready Tuberculosis data into R.

See also

Useful links:

http://www.samabbott.co.uk/getTBinR

https://github.com/seabbs/getTBinR

Report bugs at https://github.com/seabbs/getTBinR/issues