Get the WHO Data Dictionary for TB Data

Get the data dictionary for TB burden data from the WHO, see here for details. This function will first attempt to load the data from the temporary directory (tempdir). If that fails, and download_data = TRUE, it will instead download the data.

get_data_dict(
  url = NULL,
  download_data = TRUE,
  save = TRUE,
  dict_save_name = NULL,
  return = TRUE,
  verbose = FALSE,
  use_utils = FALSE,
  retry_download = TRUE
)

Arguments

url
Character string, indicating the url of the data dictionary. This argument is depreciated and will be removed from future releases.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

dict_save_name
Character string, name to save dictionary under. This argument is depreciated and will be removed from future releases. Dataset naming is now handled internally.

return
Logical, should the data dictionary be returned as a dataframe. Defaults to TRUE.

verbose
Logical, defaults to FALSE. Should additional status and progress messages be displayed.

use_utils
Logical, defaults to FALSE. Used for testing alternative data download function. When TRUE data is downloaded using read.csv.

retry_download
Logical defaults to TRUE. When TRUE, if downloading fails, the function will try repeatedly to download the data within 3 seconds, up to 3 times.

Value

The WHO TB data dictionary as a tibble with 4 variables: variable_name, dataset, code_list, definition.

See also

Examples

dict <- get_data_dict()

head(dict)

#> # A tibble: 6 x 4
#>   variable_name   dataset code_list definition                                  
#>   <chr>           <chr>   <chr>     <chr>                                       
#> 1 budget_cpp_dstb Budget  ""        Average cost of drugs budgeted per patient …
#> 2 budget_cpp_mdr  Budget  ""        Average cost of drugs budgeted per patient …
#> 3 budget_cpp_xdr  Budget  ""        Average cost of drugs budgeted per patient …
#> 4 budget_fld      Budget  ""        Budget required for drugs to treat drug-sus…
#> 5 budget_lab      Budget  ""        Budget required for laboratory infrastructu…
#> 6 budget_mdrmgt   Budget  ""        Budget required for programme costs to trea…