Get the WHO TB Burden Data

Get the TB burden data from the WHO, see here for details.

get_tb_burden(url = "https://extranet.who.int/tme/generateCSV.asp?ds=estimates",
  path = "data-raw", save = TRUE, save_name = "TB_burden",
  return = TRUE)

Arguments

url
Character string, indicating the url of the TB burden data. Default is current url.

path
Character string, indicating the folder to save the data into, defaults to data-raw.

save
Logical, should the data be saved for reuse. Defaults to TRUE.

save_name
Character string, name to save the data under. Defaults to TB_burden.

return
Logical, should the data be returned as a dataframe. Defaults to TRUE.

Value

The WHO TB burden data as a tibble.

See also

search_data_dict

Examples

tb_burden <- get_tb_burden()

head(tb_burden)

#> # A tibble: 6 x 71
#>       country  iso2  iso3 iso_numeric g_whoregion  year e_pop_num e_inc_100k
#>         <chr> <chr> <chr>       <chr>       <chr> <int>     <int>      <dbl>
#> 1 Afghanistan    AF   AFG         004         EMR  2000  20093756        190
#> 2 Afghanistan    AF   AFG         004         EMR  2001  20966463        189
#> 3 Afghanistan    AF   AFG         004         EMR  2002  21979923        189
#> 4 Afghanistan    AF   AFG         004         EMR  2003  23064851        189
#> 5 Afghanistan    AF   AFG         004         EMR  2004  24118979        189
#> 6 Afghanistan    AF   AFG         004         EMR  2005  25070798        189
#> # ... with 63 more variables: e_inc_100k_lo <dbl>, e_inc_100k_hi <dbl>,
#> #   e_inc_num <int>, e_inc_num_lo <int>, e_inc_num_hi <int>,
#> #   e_inc_num_f014 <int>, e_inc_num_f014_lo <int>, e_inc_num_f014_hi <int>,
#> #   e_inc_num_f15plus <int>, e_inc_num_f15plus_lo <int>,
#> #   e_inc_num_f15plus_hi <int>, e_inc_num_f <int>, e_inc_num_f_lo <int>,
#> #   e_inc_num_f_hi <int>, e_inc_num_m014 <int>, e_inc_num_m014_lo <int>,
#> #   e_inc_num_m014_hi <int>, e_inc_num_m15plus <int>,
#> #   e_inc_num_m15plus_lo <int>, e_inc_num_m15plus_hi <int>, e_inc_num_m <int>,
#> #   e_inc_num_m_lo <int>, e_inc_num_m_hi <int>, e_inc_num_014 <int>,
#> #   e_inc_num_014_lo <int>, e_inc_num_014_hi <int>, e_inc_num_15plus <int>,
#> #   e_inc_num_15plus_lo <int>, e_inc_num_15plus_hi <int>, e_tbhiv_prct <dbl>,
#> #   e_tbhiv_prct_lo <dbl>, e_tbhiv_prct_hi <dbl>, e_inc_tbhiv_100k <dbl>,
#> #   e_inc_tbhiv_100k_lo <dbl>, e_inc_tbhiv_100k_hi <dbl>,
#> #   e_inc_tbhiv_num <int>, e_inc_tbhiv_num_lo <int>, e_inc_tbhiv_num_hi <int>,
#> #   e_mort_exc_tbhiv_100k <dbl>, e_mort_exc_tbhiv_100k_lo <dbl>,
#> #   e_mort_exc_tbhiv_100k_hi <dbl>, e_mort_exc_tbhiv_num <int>,
#> #   e_mort_exc_tbhiv_num_lo <int>, e_mort_exc_tbhiv_num_hi <int>,
#> #   e_mort_tbhiv_100k <dbl>, e_mort_tbhiv_100k_lo <dbl>,
#> #   e_mort_tbhiv_100k_hi <dbl>, e_mort_tbhiv_num <int>,
#> #   e_mort_tbhiv_num_lo <int>, e_mort_tbhiv_num_hi <int>, e_mort_100k <dbl>,
#> #   e_mort_100k_lo <dbl>, e_mort_100k_hi <dbl>, e_mort_num <int>,
#> #   e_mort_num_lo <int>, e_mort_num_hi <int>, cfr <dbl>, cfr_lo <dbl>,
#> #   cfr_hi <dbl>, c_newinc_100k <dbl>, c_cdr <dbl>, c_cdr_lo <dbl>,
#> #   c_cdr_hi <dbl>