Summarise TB Burden

Summarise TB Burden

summarise_tb_burden(df = NULL, dict = NULL, metric = "e_inc_100k",
  metric_label = NULL, conf = c("_lo", "_hi"), years = NULL,
  samples = 1000, countries = NULL, compare_to_region = FALSE,
  compare_to_world = FALSE, custom_compare = NULL,
  compare_all_regions = FALSE, compute_rate = FALSE,
  compute_rate_with = "e_pop_num", rate_scaling = 1e+05,
  truncate_at_zero = TRUE, annual_change = FALSE, download_data = TRUE,
  save = TRUE, burden_save_name = "TB_burden",
  dict_save_name = "TB_data_dict", verbose = TRUE)

Arguments

df
Dataframe of TB burden data, as sourced by get_tb_burden. If not specified then will source the WHO TB burden data, either locally if available or directly from the WHO (if download_data = TRUE).

dict
A tibble of the data dictionary. See get_data_dict for details. If not supplied the function will attempt to load a saved version of the dictionary. If this fails and download_data = TRUE then the dictionary will be downloaded.

metric
Character string specifying the metric to plot

metric_label
Character string specifying the metric label to use.

conf
Character vector specifying the name variations to use to specify the upper and lower confidence intervals. Defaults to NULL for which no confidence intervals are used. Used by annual_change.

years
Numeric vector of years. Defaults to NULL which includes all years in the data.

samples
Numeric, the number of samples to use to generate confidence intervals (only used when conf are present)

countries
A character string specifying the countries to plot.

compare_to_region
Logical, defaults to FALSE. If TRUE all countries that share a region with those listed in countries will be plotted. Note that this will override settings for facet, unless it is set to "country".

compare_to_world
Logical, defaults to FALSE. Should a comparision be made to the metric of interests global value.

custom_compare
Logical, defaults to NULL. A named list of custom countries.

compare_all_regions
Logical, defaults to FALSE. Should all regions be compared.

compute_rate
Logical, defaults to FALSE. Should the rate be calculated using the summarised population.

compute_rate_with
Character, the variable to compute the rate with. Defaults to the using the total population ("e_pop_num").

rate_scaling
Numeric, scaling to apply to the calculate rate. Defaults to 100,000.

truncate_at_zero
Logical, defaults to TRUE. Should lower bounds be truncated at zero?

annual_change
Logical, defaults to FALSE. If TRUE then the percentage annual change is computed for the specified metric.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

burden_save_name
Character string, name to save the data under. Defaults to "TB_burden".

dict_save_name
Character string, name to save dictionary under. Defaults to TB_data_dict.

verbose
Logical, defaults to TRUE. Should additional status and progress messages be displayed.

Value

A tibble containing summarised values (with 95 stratified by area and year.

Examples

## Get summary of the e_mdr_pct_rr_new cases in 2016
summarise_tb_burden(metric = "e_mdr_pct_rr_new",
                    years = 2016,
                    samples = 100,
                    compare_all_regions = TRUE,
                    compare_to_world = TRUE,
                    verbose = TRUE)

#> Loading data from: /tmp/RtmpabTiht/TB_burden.rds

#> Loading data from: /tmp/RtmpabTiht/MDR_TB.rds

#> Joining TB burden data and MDR TB data.

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Loading data from: /tmp/RtmpabTiht/TB_data_dict.rds

#> 1 results found for your variable search for e_mdr_pct_rr_new

#> Loading data from: /tmp/RtmpabTiht/TB_burden.rds

#> Loading data from: /tmp/RtmpabTiht/MDR_TB.rds

#> Joining TB burden data and MDR TB data.

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Loading data from: /tmp/RtmpabTiht/TB_data_dict.rds

#> 1 results found for your variable search for e_mdr_pct_rr_new

#> Filtering to use only data from: 2016

#> Confidence intervals were not found using your specified conf, so defaulting to estimating
#>               only based on the point estimate.

#> # A tibble: 7 x 5
#>   area              year e_mdr_pct_rr_new e_mdr_pct_rr_new_… e_mdr_pct_rr_new_…
#>   <fct>            <int>            <dbl>              <dbl>              <dbl>
#> 1 Africa            2016             70.0               40.6               99.3
#> 2 Americas          2016             78.8               35.8              122. 
#> 3 Eastern Mediter…  2016             76.9               29.8              124. 
#> 4 Europe            2016             90.0               63.8              116. 
#> 5 South-East Asia   2016             82.5               45.1              120. 
#> 6 Western Pacific   2016             69.3               26.7              112. 
#> 7 Global            2016             77.9               38.6              117. 

## Get summary of the case fatality rate for regions and the world in 2016
## Boostrapping CI's
summarise_tb_burden(metric = "cfr",
                    years = 2016,
                    samples = 100,
                    compare_all_regions = TRUE,
                    compare_to_world = TRUE,
                    verbose = FALSE)

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> # A tibble: 7 x 5
#>   area                   year    cfr cfr_lo cfr_hi
#>   <fct>                 <int>  <dbl>  <dbl>  <dbl>
#> 1 Africa                 2016 0.239  0.0175  0.461
#> 2 Americas               2016 0.132  0       0.397
#> 3 Eastern Mediterranean  2016 0.129  0       0.393
#> 4 Europe                 2016 0.0839 0       0.176
#> 5 South-East Asia        2016 0.137  0.0286  0.245
#> 6 Western Pacific        2016 0.0888 0       0.194
#> 7 Global                 2016 0.137  0       0.361

## Get summary data for the UK, Europe and the world
## Bootstrapping CI's and calculating the incidence rate
summarise_tb_burden(metric = "e_inc_num",
                    samples = 100,
                    countries = "United Kingdom",
                    compare_to_world = TRUE,
                    compare_to_region = TRUE,
                    compute_rate = TRUE,
                    verbose = FALSE)

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Joining, by = c("Area", "Year")

#> # A tibble: 51 x 5
#>    area                                year e_inc_num e_inc_num_lo e_inc_num_hi
#>    <fct>                              <int>     <dbl>        <dbl>        <dbl>
#>  1 United Kingdom of Great Britain a…  2000      12.0         10.7         13.2
#>  2 United Kingdom of Great Britain a…  2001      11.5         10.2         12.8
#>  3 United Kingdom of Great Britain a…  2002      13.2         12.1         14.3
#>  4 United Kingdom of Great Britain a…  2003      13.3         12.0         14.6
#>  5 United Kingdom of Great Britain a…  2004      13.2         11.8         14.6
#>  6 United Kingdom of Great Britain a…  2005      15.2         13.6         16.8
#>  7 United Kingdom of Great Britain a…  2006      15.3         14.0         16.5
#>  8 United Kingdom of Great Britain a…  2007      14.7         13.0         16.3
#>  9 United Kingdom of Great Britain a…  2008      15.0         13.7         16.2
#> 10 United Kingdom of Great Britain a…  2009      14.5         12.9         16.1
#> # ... with 41 more rows