getTBinR: Get WHO Tuberculosis Data
Source: R/getTBinR.R

Facilitates easy import of analysis ready World Health Organisation Tuberculosis data and provides plotting functions for exploratory data analysis.

See also

Useful links:

http://www.samabbott.co.uk/getTBinR

https://github.com/seabbs/getTBinR

Report bugs at https://github.com/seabbs/getTBinR/issues