Generic Get Data Function

If the data is found locally in the temporary directory then this function will load the data into R. Otherwise if download_data = TRUE then the data will be retrieved from the specified URL. Data can then be saved to the temporary directory by specifying save = TRUE.

get_data(url = NULL, data_trans_fn = NULL, download_data = TRUE,
  save = TRUE, save_name = NULL, return = TRUE, verbose = TRUE,
  use_utils = FALSE, use_direct_download = FALSE,
  retry_download = TRUE)

Arguments

url
Character string, indicating the url of the data to download.

data_trans_fn
Function that takes a data.table as input and returns a single dataframe of any type. If not specified defaults to transforming the data into a tibble.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

save_name
Character string, name to save the data under. Defaults to NULL.

return
Logical, should the data be returned as a dataframe. Defaults to TRUE.

verbose
Logical, defaults to TRUE. Should additional status and progress messages be displayed.

use_utils
Logical, defaults to FALSE. Used for testing alternative data download function. When TRUE data is downloaded using read.csv.

use_direct_download
Logical, defaults to FALSE. Used for testing alternative data download function. When TRUE data is downloaded using download.file.

retry_download
Logical defaults to TRUE. When TRUE, if downloading fails, the function will try repeatedly to download the data within 3 seconds, up to 3 times.

Value

The data loaded from a local copy or downloaded from the given url as a dataframe, exact format specified by data_trans_fn

See also

Examples

tb_burden <- get_data(url = "https://extranet.who.int/tme/generateCSV.asp?ds=estimates",
save_name = "TB_burden")

#> Downloading data from: https://extranet.who.int/tme/generateCSV.asp?ds=estimates

#> Saving data to: /tmp/RtmpGUGIj7/TB_burden.rds

head(tb_burden)

#> [90m# A tibble: 6 x 47[39m
#>   country iso2  iso3  iso_numeric g_whoregion  year e_pop_num e_inc_100k
#>   [3m[90m<chr>[39m[23m   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m       [3m[90m<int>[39m[23m [3m[90m<chr>[39m[23m       [3m[90m<int>[39m[23m     [3m[90m<int>[39m[23m      [3m[90m<dbl>[39m[23m
#> [90m1[39m Afghan… AF    AFG             4 EMR          [4m2[24m000  20[4m0[24m[4m9[24m[4m3[24m756        190
#> [90m2[39m Afghan… AF    AFG             4 EMR          [4m2[24m001  20[4m9[24m[4m6[24m[4m6[24m463        189
#> [90m3[39m Afghan… AF    AFG             4 EMR          [4m2[24m002  21[4m9[24m[4m7[24m[4m9[24m923        189
#> [90m4[39m Afghan… AF    AFG             4 EMR          [4m2[24m003  23[4m0[24m[4m6[24m[4m4[24m851        189
#> [90m5[39m Afghan… AF    AFG             4 EMR          [4m2[24m004  24[4m1[24m[4m1[24m[4m8[24m979        189
#> [90m6[39m Afghan… AF    AFG             4 EMR          [4m2[24m005  25[4m0[24m[4m7[24m[4m0[24m798        189
#> [90m# ... with 39 more variables: e_inc_100k_lo [3m[90m<dbl>[90m[23m, e_inc_100k_hi [3m[90m<dbl>[90m[23m,
#> #   e_inc_num [3m[90m<int>[90m[23m, e_inc_num_lo [3m[90m<int>[90m[23m, e_inc_num_hi [3m[90m<int>[90m[23m,
#> #   e_tbhiv_prct [3m[90m<dbl>[90m[23m, e_tbhiv_prct_lo [3m[90m<dbl>[90m[23m, e_tbhiv_prct_hi [3m[90m<dbl>[90m[23m,
#> #   e_inc_tbhiv_100k [3m[90m<dbl>[90m[23m, e_inc_tbhiv_100k_lo [3m[90m<dbl>[90m[23m,
#> #   e_inc_tbhiv_100k_hi [3m[90m<dbl>[90m[23m, e_inc_tbhiv_num [3m[90m<int>[90m[23m, e_inc_tbhiv_num_lo [3m[90m<int>[90m[23m,
#> #   e_inc_tbhiv_num_hi [3m[90m<int>[90m[23m, e_mort_exc_tbhiv_100k [3m[90m<dbl>[90m[23m,
#> #   e_mort_exc_tbhiv_100k_lo [3m[90m<dbl>[90m[23m, e_mort_exc_tbhiv_100k_hi [3m[90m<dbl>[90m[23m,
#> #   e_mort_exc_tbhiv_num [3m[90m<int>[90m[23m, e_mort_exc_tbhiv_num_lo [3m[90m<int>[90m[23m,
#> #   e_mort_exc_tbhiv_num_hi [3m[90m<int>[90m[23m, e_mort_tbhiv_100k [3m[90m<dbl>[90m[23m,
#> #   e_mort_tbhiv_100k_lo [3m[90m<dbl>[90m[23m, e_mort_tbhiv_100k_hi [3m[90m<dbl>[90m[23m,
#> #   e_mort_tbhiv_num [3m[90m<int>[90m[23m, e_mort_tbhiv_num_lo [3m[90m<int>[90m[23m,
#> #   e_mort_tbhiv_num_hi [3m[90m<int>[90m[23m, e_mort_100k [3m[90m<dbl>[90m[23m, e_mort_100k_lo [3m[90m<dbl>[90m[23m,
#> #   e_mort_100k_hi [3m[90m<dbl>[90m[23m, e_mort_num [3m[90m<int>[90m[23m, e_mort_num_lo [3m[90m<int>[90m[23m,
#> #   e_mort_num_hi [3m[90m<int>[90m[23m, cfr [3m[90m<dbl>[90m[23m, cfr_lo [3m[90m<dbl>[90m[23m, cfr_hi [3m[90m<dbl>[90m[23m,
#> #   c_newinc_100k [3m[90m<dbl>[90m[23m, c_cdr [3m[90m<dbl>[90m[23m, c_cdr_lo [3m[90m<dbl>[90m[23m, c_cdr_hi [3m[90m<dbl>[90m[23m[39m