Get the WHO Data Dictionary for TB Data

Get the data dictionary for TB burden data from the WHO, see here for details.

get_data_dict(url = "https://extranet.who.int/tme/generateCSV.asp?ds=dictionary",
  path = "data-raw", save = TRUE, save_name = "TB_data_dict",
  return = TRUE)

Arguments

url
Character string, indicating the url of the data dictionary. Default is current url.

path
Character string, indicating the folder to save the data into, defaults to data-raw.

save
Logical, should the dictionary be saved for reuse. Defaults to TRUE.

save_name
Character string, name to save dictionary under. Defaults to TB_data_dict.

return
Logical, should the data dictionary be returned as a dataframe. Defaults to TRUE.

Value

The WHO TB data dictionary as a tibble with 4 variables: variable_name, dataset, code_list, definition.

See also

get_data search_data_dict

Examples

dict <- get_data_dict()

head(dict)

#> # A tibble: 6 x 4
#>              variable_name              dataset   code_list
#>                      <chr>                <chr>       <chr>
#> 1                      bmu Community engagement            
#> 2       bmu_community_impl Community engagement            
#> 3             bmu_ref_data Community engagement            
#> 4       bmu_rxsupport_data Community engagement            
#> 5   bmu_rxsupport_data_coh Community engagement            
#> 6 community_data_available Community engagement 0=No; 1=Yes
#> # ... with 1 more variables: definition <chr>