Get the WHO Data Dictionary for TB Data

Get the data dictionary for TB burden data from the WHO, see here for details. This function will first attempt to load the data from the temporary directory (tempdir). If that fails, and download_data = TRUE, it will instead download the data.

get_data_dict(url = "https://extranet.who.int/tme/generateCSV.asp?ds=dictionary",
  download_data = TRUE, save = TRUE, dict_save_name = "TB_data_dict",
  return = TRUE, verbose = TRUE, use_utils = FALSE,
  retry_download = TRUE)

Arguments

url
Character string, indicating the url of the data dictionary. Default is current url.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

dict_save_name
Character string, name to save dictionary under. Defaults to TB_data_dict.

return
Logical, should the data dictionary be returned as a dataframe. Defaults to TRUE.

verbose
Logical, defaults to TRUE. Should additional status and progress messages be displayed.

use_utils
Logical, defaults to FALSE. Used for testing alternative data download function. When TRUE data is downloaded using read.csv.

retry_download
Logical defaults to TRUE. When TRUE, if downloading fails, the function will try repeatedly to download the data within 3 seconds, up to 3 times.

Value

The WHO TB data dictionary as a tibble with 4 variables: variable_name, dataset, code_list, definition.

See also

Examples

dict <- get_data_dict()

#> Downloading data from: https://extranet.who.int/tme/generateCSV.asp?ds=dictionary

#> Saving data to: /tmp/Rtmp84wfNu/TB_data_dict.rds

head(dict)

#> [90m# A tibble: 6 x 4[39m
#>   variable_name  dataset code_list definition                                  
#>   [3m[90m<chr>[39m[23m          [3m[90m<chr>[39m[23m   [3m[90m<chr>[39m[23m     [3m[90m<chr>[39m[23m                                       
#> [90m1[39m budget_cpp_ds… Budget  [90m"[39m[90m"[39m        Average cost of drugs budgeted per patient …
#> [90m2[39m budget_cpp_mdr Budget  [90m"[39m[90m"[39m        Average cost of drugs budgeted per patient …
#> [90m3[39m budget_cpp_xdr Budget  [90m"[39m[90m"[39m        Average cost of drugs budgeted per patient …
#> [90m4[39m budget_fld     Budget  [90m"[39m[90m"[39m        Budget required for drugs to treat drug-sus…
#> [90m5[39m budget_lab     Budget  [90m"[39m[90m"[39m        Budget required for laboratory infrastructu…
#> [90m6[39m budget_mdrmgt  Budget  [90m"[39m[90m"[39m        Budget required for programme costs to trea…