Get the WHO TB Burden Data

Get the TB burden data from the WHO, see here for details. This function will first attempt to load the data from the temporary directory (tempdir). If that fails, and download_data = TRUE, it will instead download the data.

get_tb_burden(url = "https://extranet.who.int/tme/generateCSV.asp?ds=estimates",
  download_data = TRUE, save = TRUE, burden_save_name = "TB_burden",
  return = TRUE, verbose = TRUE, use_utils = FALSE,
  retry_download = TRUE)

Arguments

url
Character string, indicating the url of the TB burden data. Default is current url.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

burden_save_name
Character string, name to save the data under. Defaults to TB_burden.

return
Logical, should the data be returned as a dataframe. Defaults to TRUE.

verbose
Logical, defaults to TRUE. Should additional status and progress messages be displayed.

use_utils
Logical, defaults to FALSE. Used for testing alternative data download function. When TRUE data is downloaded using read.csv.

retry_download
Logical defaults to TRUE. When TRUE, if downloading fails, the function will try repeatedly to download the data within 5 seconds, up to 5 times.

Value

The WHO TB burden data as a tibble.

See also

get_data search_data_dict

Examples

tb_burden <- get_tb_burden()

#> Loading data from: /tmp/RtmplXxb0Q/TB_burden.rds

head(tb_burden)

#> # A tibble: 6 x 71
#>   country   iso2  iso3  iso_numeric g_whoregion       year e_pop_num e_inc_100k
#>   <chr>     <chr> <chr>       <int> <chr>            <int>     <int>      <dbl>
#> 1 Afghanis… AF    AFG             4 Eastern Mediter…  2000  20093756        190
#> 2 Afghanis… AF    AFG             4 Eastern Mediter…  2001  20966463        189
#> 3 Afghanis… AF    AFG             4 Eastern Mediter…  2002  21979923        189
#> 4 Afghanis… AF    AFG             4 Eastern Mediter…  2003  23064851        189
#> 5 Afghanis… AF    AFG             4 Eastern Mediter…  2004  24118979        189
#> 6 Afghanis… AF    AFG             4 Eastern Mediter…  2005  25070798        189
#> # ... with 63 more variables: e_inc_100k_lo <dbl>, e_inc_100k_hi <dbl>,
#> #   e_inc_num <int>, e_inc_num_lo <int>, e_inc_num_hi <int>,
#> #   e_inc_num_f014 <int>, e_inc_num_f014_lo <int>, e_inc_num_f014_hi <int>,
#> #   e_inc_num_f15plus <int>, e_inc_num_f15plus_lo <int>,
#> #   e_inc_num_f15plus_hi <int>, e_inc_num_f <int>, e_inc_num_f_lo <int>,
#> #   e_inc_num_f_hi <int>, e_inc_num_m014 <int>, e_inc_num_m014_lo <int>,
#> #   e_inc_num_m014_hi <int>, e_inc_num_m15plus <int>,
#> #   e_inc_num_m15plus_lo <int>, e_inc_num_m15plus_hi <int>, e_inc_num_m <int>,
#> #   e_inc_num_m_lo <int>, e_inc_num_m_hi <int>, e_inc_num_014 <int>,
#> #   e_inc_num_014_lo <int>, e_inc_num_014_hi <int>, e_inc_num_15plus <int>,
#> #   e_inc_num_15plus_lo <int>, e_inc_num_15plus_hi <int>, e_tbhiv_prct <dbl>,
#> #   e_tbhiv_prct_lo <dbl>, e_tbhiv_prct_hi <dbl>, e_inc_tbhiv_100k <dbl>,
#> #   e_inc_tbhiv_100k_lo <dbl>, e_inc_tbhiv_100k_hi <dbl>,
#> #   e_inc_tbhiv_num <int>, e_inc_tbhiv_num_lo <int>, e_inc_tbhiv_num_hi <int>,
#> #   e_mort_exc_tbhiv_100k <dbl>, e_mort_exc_tbhiv_100k_lo <dbl>,
#> #   e_mort_exc_tbhiv_100k_hi <dbl>, e_mort_exc_tbhiv_num <int>,
#> #   e_mort_exc_tbhiv_num_lo <int>, e_mort_exc_tbhiv_num_hi <int>,
#> #   e_mort_tbhiv_100k <dbl>, e_mort_tbhiv_100k_lo <dbl>,
#> #   e_mort_tbhiv_100k_hi <dbl>, e_mort_tbhiv_num <int>,
#> #   e_mort_tbhiv_num_lo <int>, e_mort_tbhiv_num_hi <int>, e_mort_100k <dbl>,
#> #   e_mort_100k_lo <dbl>, e_mort_100k_hi <dbl>, e_mort_num <int>,
#> #   e_mort_num_lo <int>, e_mort_num_hi <int>, cfr <dbl>, cfr_lo <dbl>,
#> #   cfr_hi <dbl>, c_newinc_100k <dbl>, c_cdr <dbl>, c_cdr_lo <dbl>,
#> #   c_cdr_hi <dbl>