Generic Function to Prepare TB Burden Data for Plotting

This function is used internally by plot_tb_burden and plot_tb_burden_overview to prepare data for plotting.

prepare_df_plot(df = NULL, dict = NULL, metric = "e_inc_100k",
  metric_label = NULL, countries = NULL, compare_to_region = FALSE,
  facet = NULL, download_data = FALSE, save = FALSE,
  burden_save_name = "TB_burden", dict_save_name = "TB_data_dict",
  verbose = TRUE, ...)

Arguments

df
Dataframe of TB burden data, as sourced by get_tb_burden. If not specified then will source the WHO TB burden data, either locally if available or directly from the WHO (if download_data = TRUE).

dict
A tibble of the data dictionary. See get_data_dict for details. If not supplied the function will attempt to load a saved version of the dictionary. If this fails and download_data = TRUE then the dictionary will be downloaded.

metric
Character string specifying the metric to plot

metric_label
Character string specifying the metric label to use.

countries
A character string specifying the countries to plot.

compare_to_region
Logical, defaults to FALSE. If TRUE all countries that share a region with those listed in countries will be plotted. Note that this will override settings for facet, unless it is set to "country".

facet
Character string, the name of the variable to facet by.

download_data
Logical, defaults to FALSE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to FALSE. If TRUE then the data is saved to the temporary directory specified by tempdir.

burden_save_name
Character string, name to save the data under. Defaults to TB_burden.

dict_save_name
Character string, name to save dictionary under. Defaults to TB_data_dict.

verbose
Logical, defaults to TRUE. Should additional status and progress messages be displayed.

...
Additional parameters to pass to get_tb_burden.

Value

A list containing 3 elements, the dataframe to plot, the facet to use and the label to assign to the metric axis.

See also

plot_tb_burden plot_tb_burden_overview

Examples

prepare_df_plot(countries = "Guinea", download_data = TRUE, save = TRUE)

#> Loading data from: /tmp/RtmpEjKCaH/TB_burden.rds

#> Loading data from: /tmp/RtmpEjKCaH/TB_data_dict.rds

#> 1 results found for your variable search for e_inc_100k

#> $df
#> # A tibble: 17 x 72
#>    count… iso2  iso3  iso_n… g_who…  year e_pop_… e_in… e_in… e_in… e_in… e_in…
#>    <fctr> <chr> <chr> <chr>  <chr>  <int>   <int> <dbl> <dbl> <dbl> <int> <int>
#>  1 Guinea GN    GIN   324    Africa  2000  8.81e⁶   228   148   326 20000 13000
#>  2 Guinea GN    GIN   324    Africa  2001  8.97e⁶   224   145   320 20000 13000
#>  3 Guinea GN    GIN   324    Africa  2002  9.14e⁶   221   143   315 20000 13000
#>  4 Guinea GN    GIN   324    Africa  2003  9.31e⁶   218   141   311 20000 13000
#>  5 Guinea GN    GIN   324    Africa  2004  9.49e⁶   214   139   306 20000 13000
#>  6 Guinea GN    GIN   324    Africa  2005  9.68e⁶   210   136   300 20000 13000
#>  7 Guinea GN    GIN   324    Africa  2006  9.88e⁶   206   133   294 20000 13000
#>  8 Guinea GN    GIN   324    Africa  2007  1.01e⁷   200   130   286 20000 13000
#>  9 Guinea GN    GIN   324    Africa  2008  1.03e⁷   195   126   279 20000 13000
#> 10 Guinea GN    GIN   324    Africa  2009  1.06e⁷   190   123   272 20000 13000
#> 11 Guinea GN    GIN   324    Africa  2010  1.08e⁷   186   120   266 20000 13000
#> 12 Guinea GN    GIN   324    Africa  2011  1.10e⁷   183   118   261 20000 13000
#> 13 Guinea GN    GIN   324    Africa  2012  1.13e⁷   180   116   257 20000 13000
#> 14 Guinea GN    GIN   324    Africa  2013  1.15e⁷   177   115   253 20000 13000
#> 15 Guinea GN    GIN   324    Africa  2014  1.18e⁷   177   114   253 21000 14000
#> 16 Guinea GN    GIN   324    Africa  2015  1.21e⁷   177   114   252 21000 14000
#> 17 Guinea GN    GIN   324    Africa  2016  1.24e⁷   176   114   252 22000 14000
#> # ... with 60 more variables: e_inc_num_hi <int>, e_inc_num_f014 <int>,
#> #   e_inc_num_f014_lo <int>, e_inc_num_f014_hi <int>, e_inc_num_f15plus <int>,
#> #   e_inc_num_f15plus_lo <int>, e_inc_num_f15plus_hi <int>, e_inc_num_f <int>,
#> #   e_inc_num_f_lo <int>, e_inc_num_f_hi <int>, e_inc_num_m014 <int>,
#> #   e_inc_num_m014_lo <int>, e_inc_num_m014_hi <int>, e_inc_num_m15plus <int>,
#> #   e_inc_num_m15plus_lo <int>, e_inc_num_m15plus_hi <int>, e_inc_num_m <int>,
#> #   e_inc_num_m_lo <int>, e_inc_num_m_hi <int>, e_inc_num_014 <int>,
#> #   e_inc_num_014_lo <int>, e_inc_num_014_hi <int>, e_inc_num_15plus <int>,
#> #   e_inc_num_15plus_lo <int>, e_inc_num_15plus_hi <int>, e_tbhiv_prct <dbl>,
#> #   e_tbhiv_prct_lo <dbl>, e_tbhiv_prct_hi <dbl>, e_inc_tbhiv_100k <dbl>,
#> #   e_inc_tbhiv_100k_lo <dbl>, e_inc_tbhiv_100k_hi <dbl>,
#> #   e_inc_tbhiv_num <int>, e_inc_tbhiv_num_lo <int>, e_inc_tbhiv_num_hi <int>,
#> #   e_mort_exc_tbhiv_100k <dbl>, e_mort_exc_tbhiv_100k_lo <dbl>,
#> #   e_mort_exc_tbhiv_100k_hi <dbl>, e_mort_exc_tbhiv_num <int>,
#> #   e_mort_exc_tbhiv_num_lo <int>, e_mort_exc_tbhiv_num_hi <int>,
#> #   e_mort_tbhiv_100k <dbl>, e_mort_tbhiv_100k_lo <dbl>,
#> #   e_mort_tbhiv_100k_hi <dbl>, e_mort_tbhiv_num <int>,
#> #   e_mort_tbhiv_num_lo <int>, e_mort_tbhiv_num_hi <int>, e_mort_100k <dbl>,
#> #   e_mort_100k_lo <dbl>, e_mort_100k_hi <dbl>, e_mort_num <int>,
#> #   e_mort_num_lo <int>, e_mort_num_hi <int>, cfr <dbl>, cfr_lo <dbl>,
#> #   cfr_hi <dbl>, c_newinc_100k <dbl>, c_cdr <dbl>, c_cdr_lo <dbl>,
#> #   c_cdr_hi <dbl>, Year <int>
#> 
#> $facet
#> NULL
#> 
#> $metric_label
#> [1] "Estimated incidence (all forms) per 100 000 population"
#>