Generic Function to Prepare TB Burden Data for Plotting

This function is used internally by plot_tb_burden and plot_tb_burden_overview to prepare data for plotting.

prepare_df_plot(df = NULL, dict = NULL, metric = "e_inc_100k",
  conf = NULL, metric_label = NULL, countries = NULL, years = NULL,
  compare_to_region = FALSE, facet = NULL, annual_change = FALSE,
  trans = "identity", download_data = TRUE, save = TRUE,
  burden_save_name = "TB_burden", dict_save_name = "TB_data_dict",
  verbose = TRUE, ...)

Arguments

df
Dataframe of TB burden data, as sourced by get_tb_burden. If not specified then will source the WHO TB burden data, either locally if available or directly from the WHO (if download_data = TRUE).

dict
A tibble of the data dictionary. See get_data_dict for details. If not supplied the function will attempt to load a saved version of the dictionary. If this fails and download_data = TRUE then the dictionary will be downloaded.

metric
Character string specifying the metric to plot

conf
Character vector specifying the name variations to use to specify the upper and lower confidence intervals. Defaults to NULL for which no confidence intervals are used. Used by annual_change.

metric_label
Character string specifying the metric label to use.

countries
A character string specifying the countries to plot.

years
Numeric vector of years. Defaults to NULL which includes all years in the data.

compare_to_region
Logical, defaults to FALSE. If TRUE all countries that share a region with those listed in countries will be plotted. Note that this will override settings for facet, unless it is set to "country".

facet
Character string, the name of the variable to facet by.

annual_change
Logical, defaults to FALSE. If TRUE then the percentage annual change is computed for the specified metric.

trans
A character string specifying the transform to use on the specified metric. Defaults to no transform ("identity"). Other options include log scaling ("log") and log base 10 scaling ("log10"). For a complete list of options see ggplot2::continous_scale.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

burden_save_name
Character string, name to save the data under. Defaults to "TB_burden".

dict_save_name
Character string, name to save dictionary under. Defaults to TB_data_dict.

verbose
Logical, defaults to TRUE. Should additional status and progress messages be displayed.

...
Additional parameters to pass to get_tb_burden.

Value

A list containing 3 elements, the dataframe to plot, the facet to use and the label to assign to the metric axis.

See also

Examples

prepare_df_plot(countries = "Guinea")

#> Loading data from: /tmp/Rtmpj5bgNm/TB_burden.rds

#> Loading data from: /tmp/Rtmpj5bgNm/MDR_TB.rds

#> Joining TB burden data and MDR TB data.

#> Joining, by = c("country", "iso2", "iso3", "iso_numeric", "year")

#> Loading data from: /tmp/Rtmpj5bgNm/TB_data_dict.rds

#> 1 results found for your variable search for e_inc_100k

#> $df
#> # A tibble: 18 x 70
#>    country iso2  iso3  iso_numeric g_whoregion  year e_pop_num e_inc_100k
#>    <fct>   <chr> <chr>       <int> <chr>       <int>     <int>      <dbl>
#>  1 Guinea  GN    GIN           324 AFR          2000   8808546        228
#>  2 Guinea  GN    GIN           324 AFR          2001   8971139        224
#>  3 Guinea  GN    GIN           324 AFR          2002   9137345        221
#>  4 Guinea  GN    GIN           324 AFR          2003   9309848        218
#>  5 Guinea  GN    GIN           324 AFR          2004   9490229        214
#>  6 Guinea  GN    GIN           324 AFR          2005   9679745        210
#>  7 Guinea  GN    GIN           324 AFR          2006   9881428        206
#>  8 Guinea  GN    GIN           324 AFR          2007  10096727        200
#>  9 Guinea  GN    GIN           324 AFR          2008  10323142        195
#> 10 Guinea  GN    GIN           324 AFR          2009  10556524        190
#> 11 Guinea  GN    GIN           324 AFR          2010  10794170        186
#> 12 Guinea  GN    GIN           324 AFR          2011  11035170        183
#> 13 Guinea  GN    GIN           324 AFR          2012  11281469        180
#> 14 Guinea  GN    GIN           324 AFR          2013  11536615        177
#> 15 Guinea  GN    GIN           324 AFR          2014  11805509        177
#> 16 Guinea  GN    GIN           324 AFR          2015  12091533        177
#> 17 Guinea  GN    GIN           324 AFR          2016  12395924        176
#> 18 Guinea  GN    GIN           324 AFR          2017  12717176        176
#> # … with 62 more variables: e_inc_100k_lo <dbl>, e_inc_100k_hi <dbl>,
#> #   e_inc_num <int>, e_inc_num_lo <int>, e_inc_num_hi <int>,
#> #   e_tbhiv_prct <dbl>, e_tbhiv_prct_lo <dbl>, e_tbhiv_prct_hi <dbl>,
#> #   e_inc_tbhiv_100k <dbl>, e_inc_tbhiv_100k_lo <dbl>,
#> #   e_inc_tbhiv_100k_hi <dbl>, e_inc_tbhiv_num <int>, e_inc_tbhiv_num_lo <int>,
#> #   e_inc_tbhiv_num_hi <int>, e_mort_exc_tbhiv_100k <dbl>,
#> #   e_mort_exc_tbhiv_100k_lo <dbl>, e_mort_exc_tbhiv_100k_hi <dbl>,
#> #   e_mort_exc_tbhiv_num <int>, e_mort_exc_tbhiv_num_lo <int>,
#> #   e_mort_exc_tbhiv_num_hi <int>, e_mort_tbhiv_100k <dbl>,
#> #   e_mort_tbhiv_100k_lo <dbl>, e_mort_tbhiv_100k_hi <dbl>,
#> #   e_mort_tbhiv_num <int>, e_mort_tbhiv_num_lo <int>,
#> #   e_mort_tbhiv_num_hi <int>, e_mort_100k <dbl>, e_mort_100k_lo <dbl>,
#> #   e_mort_100k_hi <dbl>, e_mort_num <int>, e_mort_num_lo <int>,
#> #   e_mort_num_hi <int>, cfr <dbl>, cfr_lo <dbl>, cfr_hi <dbl>,
#> #   c_newinc_100k <dbl>, c_cdr <dbl>, c_cdr_lo <dbl>, c_cdr_hi <dbl>,
#> #   source_rr_new <chr>, source_drs_coverage_new <chr>,
#> #   source_drs_year_new <int>, e_rr_pct_new <dbl>, e_rr_pct_new_lo <dbl>,
#> #   e_rr_pct_new_hi <dbl>, e_mdr_pct_rr_new <int>, source_rr_ret <chr>,
#> #   source_drs_coverage_ret <chr>, source_drs_year_ret <int>,
#> #   e_rr_pct_ret <dbl>, e_rr_pct_ret_lo <dbl>, e_rr_pct_ret_hi <dbl>,
#> #   e_mdr_pct_rr_ret <int>, e_inc_rr_num <int>, e_inc_rr_num_lo <int>,
#> #   e_inc_rr_num_hi <int>, e_mdr_pct_rr <int>, e_rr_in_notified_pulm <int>,
#> #   e_rr_in_notified_pulm_lo <int>, e_rr_in_notified_pulm_hi <int>, `Estimated
#> #   incidence (all forms) per 100 000 population` <dbl>, Year <int>
#> 
#> $facet
#> NULL
#> 
#> $metric_label
#> [1] "Estimated incidence (all forms) per 100 000 population"
#>