Search the WHO TB Data Dictionary by Variable Name.

Searches the WHO data dictionary for TB burden data. When run for the first time it will download the data dictionary, if download_data = TRUE, and save it into the temporary

search_data_dict(var = NULL, def = NULL, dict = NULL,
  download_data = TRUE, save = TRUE, dict_save_name = "TB_data_dict",
  verbose = TRUE, ...)

Arguments

var
A character vector of variable names.

def
A character vector of terms to use to search the variable definitions for partial matches

dict
A tibble of the data dictionary. See get_data_dict for details. If not supplied the function will attempt to load a saved version of the dictionary. If this fails and download_data = TRUE then the dictionary will be downloaded.

download_data
Logical, defaults to TRUE. If not found locally should the data be downloaded from the specified URL?

save
Logical, should the data be saved for reuse during the current R session. Defaults to TRUE. If TRUE then the data is saved to the temporary directory specified by tempdir.

dict_save_name
Character string, name to save dictionary under. Defaults to TB_data_dict.

verbose
Logical, defaults to TRUE. Should search information be returned.

...
Additional parameters to pass to get_data_dict.

Value

A tibble containing the information in the data dictionary matching the variables searched for.

See also

Examples

## Search for a known variable
## Download and save the dictionary if it is not available locally
search_data_dict(var = "country")

#> Loading data from: /tmp/RtmpJ3NOwP/TB_data_dict.rds

#> 1 results found for your variable search for country

#> [90m# A tibble: 1 x 4[39m
#>   variable_name dataset                code_list definition               
#>   [3m[90m<chr>[39m[23m         [3m[90m<chr>[39m[23m                  [3m[90m<chr>[39m[23m     [3m[90m<chr>[39m[23m                    
#> [90m1[39m country       Country identification [90m"[39m[90m"[39m        Country or territory name

## Search for all variables mentioning mortality in their definition
search_data_dict(def = "mortality")

#> Loading data from: /tmp/RtmpJ3NOwP/TB_data_dict.rds

#> 9 results found for your definition search for mortality

#> [90m# A tibble: 9 x 4[39m
#>   variable_name      dataset  code_list definition                             
#>   [3m[90m<chr>[39m[23m              [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m     [3m[90m<chr>[39m[23m                                  
#> [90m1[39m e_mort_100k        Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m2[39m e_mort_100k_hi     Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m3[39m e_mort_100k_lo     Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m4[39m e_mort_exc_tbhiv_… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m5[39m e_mort_exc_tbhiv_… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m6[39m e_mort_exc_tbhiv_… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m7[39m e_mort_tbhiv_100k  Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases who ar…
#> [90m8[39m e_mort_tbhiv_100k… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases who ar…
#> [90m9[39m e_mort_tbhiv_100k… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases who ar…

## Search for both a known variable and for mortality being mentioned in there definition
## Duplicate entries will be omitted.
search_data_dict(var = "e_mort_exc_tbhiv_100k", def = "mortality")

#> Loading data from: /tmp/RtmpJ3NOwP/TB_data_dict.rds

#> 1 results found for your variable search for e_mort_exc_tbhiv_100k

#> 9 results found for your definition search for mortality

#> [90m# A tibble: 9 x 4[39m
#>   variable_name      dataset  code_list definition                             
#>   [3m[90m<chr>[39m[23m              [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m     [3m[90m<chr>[39m[23m                                  
#> [90m1[39m e_mort_exc_tbhiv_… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m2[39m e_mort_100k        Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m3[39m e_mort_100k_hi     Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m4[39m e_mort_100k_lo     Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m5[39m e_mort_exc_tbhiv_… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m6[39m e_mort_exc_tbhiv_… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases (all f…
#> [90m7[39m e_mort_tbhiv_100k  Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases who ar…
#> [90m8[39m e_mort_tbhiv_100k… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases who ar…
#> [90m9[39m e_mort_tbhiv_100k… Estimat… [90m"[39m[90m"[39m        Estimated mortality of TB cases who ar…